Opciones dashboard agregadas

+++ src/app/dashboard/components/cursos/add-student/add-student.component.ts
@@ -1,13 +1,14 @@
 import { Component, OnInit, ViewChild, AfterViewInit } from '@angular/core';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import {MatDialog} from '@angular/material/dialog';
 import { UsersService } from 'src/app/core/services/users/users.service';
 import { MatTableDataSource } from '@angular/material/table';
 import { LessonsService } from 'src/app/core/services/lessons/lessons.service';
 import { MatricularComponent } from './matricular/matricular.component';
 import { CourseService } from 'src/app/core/services/courses/course.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-add-student',
@@ -24,6 +25,7 @@ export class AddStudentComponent implements OnInit, AfterViewInit  {
 
 
   idCurso:string;
+  course;
   students:any;
   enrolledStudents: any[];
   lessons:any[];
@@ -34,7 +36,8 @@ export class AddStudentComponent implements OnInit, AfterViewInit  {
     private lessonService: LessonsService,
     private userService: UsersService,
     private activatedRoute: ActivatedRoute,
-    public dialog: MatDialog
+    public dialog: MatDialog,
+    private router: Router,
   )
   {
     this.content = [];
@@ -42,6 +45,10 @@ export class AddStudentComponent implements OnInit, AfterViewInit  {
     this.idCurso = this.activatedRoute.snapshot.params.idCurso;
     this.getLessons();
     this.obtenerListaEstudianteMatriculados();
+  }
+
+  ngOnInit(): void {
+    this.getCourse();
   }
 
   ngAfterViewInit(): void {
@@ -56,6 +63,16 @@ export class AddStudentComponent implements OnInit, AfterViewInit  {
       width: '80%',
       data:this.idCurso
     });
+  }
+
+  getCourse() {
+    let courseReceived = this.courseService.detailCourse(this.idCurso)
+      .valueChanges()
+      .subscribe(c => {
+        this.course = c;
+        console.log(this.course);
+        courseReceived.unsubscribe();
+      });
   }
 
   getLessons()
@@ -99,18 +116,49 @@ export class AddStudentComponent implements OnInit, AfterViewInit  {
     .subscribe(estudiantesMatriculados => this.dataSource.data = estudiantesMatriculados);
   }
 
-  ngOnInit(): void {
-  }
-
   applyFilter(filterValue: string): void {
     this.dataSource.filter = filterValue.trim().toLowerCase();
   }
 
   matricularEstudiante(estudiante:any)
   {
-    const contenidoCurso = {}
-    this.userService.verificarMatriculaDelEstudiante(estudiante.id, this.idCurso, contenidoCurso);
-    this.courseService.deleteEnrolledStudent(this.idCurso, estudiante.id);
+    Swal.fire({
+      title: '¿Esta seguro?',
+      text: 'Esta acción desmatricula de este curso al estudiante seleccionado! ¿Esta seguro?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, estoy seguro!'
+    })
+    .then((result) => {
+      if (result.value) {
+        const contenidoCurso = {}
+        this.userService.verificarMatriculaDelEstudiante(estudiante.id, this.idCurso, contenidoCurso);
+        this.courseService.deleteEnrolledStudent(this.idCurso, estudiante.id)
+        .then(() => {
+          Swal.fire({
+            icon: 'success',
+            title: 'Exito!',
+            text: 'Estudiante desmatriculado exitosamente',
+            confirmButtonText: 'cerrar',
+        });
+    })
+    .catch((error) => {
+      Swal.fire({
+        icon: 'error',
+        title: 'error',
+        text: 'Ocurrió un error' + error,
+        confirmButtonText: 'cerrar',
+            });
+        });
+      }
+    })
+    .catch(error => console.log(error));
+  }
+
+  goBack() {
+    this.router.navigate([`dashboard/cursos`]);
   }
 
 }
